Update the document content at each change
Since the client only provides the full content when opening the
document, we need to apply each change in order to have an up-to-date
version of the content.

This comment implements changes reconciliation and moves the parsing of
the content at the time it is needed, not in the stored data.