Refactor custom initialization
This commit changes the response to an INITIALIZE request. Before, we
overrode LanguageServerProtocol.lsp_initialize in our SaltLspProto
class. As a result we had to manually call it. That was not a good idea,
because it was decorated, pygls to called our initialize routine, which
called the original lsp_initialize in a loop, until a RecursionError
happened.

The new approach is to leave lsp_initialize as it is and just change the
one attribute we want to override in its own function.